requirements.txt
streamlit
psycopg2-binary
python-dotenv
openai>=1.0.0